add bento_archive_path, bento_service pytest fixture
