+++ tests/cli/test_cli.py
@@ -3,13 +3,9 @@ import tempfile
 import sys
 import json
 from click.testing import CliRunner
-from tests.utils import generate_fake_dataframe_model
 
-try:
-    from bentoml.cli import create_bentoml_cli
-except ImportError:
-    sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-    from bentoml.cli import create_bentoml_cli
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+from bentoml.cli import create_bentoml_cli  # noqa: E402
 
 
 def generate_test_input_file():
@@ -23,14 +19,13 @@ def generate_test_input_file():
     return file_path
 
 
-def test_run_command_with_input_file():
-    saved_path = generate_fake_dataframe_model()
+def test_run_command_with_input_file(bento_archive_path):
     input_path = generate_test_input_file()
     runner = CliRunner()
 
     cli = create_bentoml_cli()
     run_cmd = cli.commands["<API_NAME>"]
-    result = runner.invoke(run_cmd, ['predict', saved_path, '--input', input_path])
+    result = runner.invoke(run_cmd, ['predict', bento_archive_path, '--input', input_path])
 
     assert result.exit_code == 0
     result_json = json.loads(result.output)
