+++ tests/server/test_model_api_server.py
@@ -2,27 +2,12 @@ import os
 import json
 import sys
 
-try:
-    import bentoml
-    from bentoml.server import BentoAPIServer
-    from tests.utils import generate_fake_dataframe_model
-except ImportError:
-    sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
-    import bentoml
-    from bentoml.server import BentoAPIServer
-    from tests.utils import generate_fake_dataframe_model
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
+from bentoml.server import BentoAPIServer  # noqa: E402
 
 
-def create_rest_server():
-    saved_path = generate_fake_dataframe_model()
-    model_service = bentoml.load(saved_path)
-
-    rest_server = BentoAPIServer(model_service)
-    return rest_server
-
-
-def test_api_function_route():
-    rest_server = create_rest_server()
+def test_api_function_route(bento_service):
+    rest_server = BentoAPIServer(bento_service)
     test_client = rest_server.app.test_client()
 
     index_list = []
